requirements.txt
absl-py==0.9.0
altgraph==0.17
asgiref==3.2.7
astunparse==1.6.3
browser-cookie3==0.11.1
browsercookie==0.7.7
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
deepface==0.0.26
Django==3.0.7
dlib @ file:///E:/Projects/Proctoring/Include/proctoring/dlib-19.8.1-cp36-cp36m-win_amd64.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
fbs==0.8.8
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gast==0.3.3
gdown==3.11.1
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.6
image==1.5.32
importlib-metadata==1.6.1
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.2.1
levelpy==0.3.2
lz4==3.1.0
macholib==1.14
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.4
pbkdf2==1.3
pefile==2019.4.18
Pillow==7.1.2
protobuf==3.12.2
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pycryptodome==3.9.8
PyInstaller==3.4
PyQt5==5.15.0
PyQt5-sip==12.8.0
PyQtWebEngine==5.15.0
PySocks==1.7.1
python-dateutil==2.8.1
pythonnet==2.5.0
pytz==2020.1
pywebview==3.2
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.18.4
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
selenium==3.141.0
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.46.1
urllib3==1.22
Werkzeug==1.0.1
WMI==1.5.1
wrapt==1.12.1
zipp==3.1.0
